+++ app/Http/Controllers/AdminUtama/DudiController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\AdminUtama;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Dudi;
+use App\Models\DudiJurusan;
 
 class DudiController extends Controller
 {
@@ -16,6 +17,7 @@ class DudiController extends Controller
 
     public function store(Request $request)
     {
+        // Cek apakah nama DUDI sudah digunakan sebelumnya
         if (Dudi::where('nama_dudi', $request->nama_dudi)->exists()) {
             return response()->json([
                 'success' => false,
@@ -28,13 +30,14 @@ class DudiController extends Controller
             'alamat' => 'required|string|max:255',
             'bidang_usaha' => 'required|string|max:255',
             'telp' => 'required|string|max:255',
-            'email' => 'required|string|max:255',
+            'email' => 'required|email|max:255',
         ],[
             'nama_dudi.required' => 'Nama DUDI harus diisi',
             'alamat.required' => 'Alamat DUDI harus diisi',
             'bidang_usaha.required' => 'Bidang usaha harus diisi',
             'telp.required' => 'Nomor telepon DUDI harus diisi',
             'email.required' => 'Email DUDI harus diisi',
+            'email.email' => 'Format email tidak valid',
         ]);
 
         Dudi::create($request->all());
@@ -85,6 +88,17 @@ class DudiController extends Controller
     public function destroy($id)
     {
         $dudi = Dudi::findOrFail($id);
+
+        // Cek relasi ke tabel lain
+        if (
+            DudiJurusan::where('dudi_id', $dudi->id)->exists() 
+        ) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Lokasi tidak bisa dihapus karena digunakan di data lain'
+            ], 400);
+        }
+
         $dudi->delete();
 
         return response()->json(['success' => true, 'message' => 'Lokasi berhasil dihapus']);
